Remove yanky.nvim and nvim-orgmode plugins
Remove yanky.nvim because I feel it will be better to do the same with
built-in functionality instead of using a plugin. Otherwise really
enjoyed using it.
Remove nvim-orgmode. I did use it a lot, but I feel it does not offer
much compared to Markdown which I can use with other tools.